[commons] refactoring: 'user' module moved to 'api'
